+++ src/Goods/Household/Hammer.java
@@ -2,7 +2,7 @@ package Goods.Household;
 
 public class Hammer extends Household
 {
-    public Hammer(int price, long serialNumber)
+    public Hammer(int price, int serialNumber)
     {
         super("Hammer", price, serialNumber);
     }
