+++ src/Goods/Household/Screwdriver.java
@@ -2,7 +2,7 @@ package Goods.Household;
 
 public class Screwdriver extends Household
 {
-    public Screwdriver(int price, long serialNumber)
+    public Screwdriver(int price, int serialNumber)
     {
         super("Screwdriver", price, serialNumber);
     }
