+++ src/Goods/Household/Shovel.java
@@ -2,7 +2,7 @@ package Goods.Household;
 
 public class Shovel extends Household
 {
-    public Shovel(int price, long serialNumber)
+    public Shovel(int price, int serialNumber)
     {
         super("Shovel", price, serialNumber);
     }
